% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdpop_setup.R
\name{cdpop_setup}
\alias{cdpop_setup}
\title{Execute to run CDPOP from R}
\usage{
cdpop_setup()
}
\description{
Function to format CDPOP to run with R
}
\examples{
## Not run:
## cdpop_setup()

## End (Not run)
}
\author{
Bill Peterman \href{mailto:Peterman.73@osu.edu}{Peterman.73@osu.edu}
}
