% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDPOPr-package.R
\docType{package}
\name{CDPOPr-package}
\alias{CDPOPr}
\alias{CDPOPr-package}
\title{CDPOPr: Run CDPOP from R}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Bill Peterman \email{Peterman.73@osu.edu} (\href{https://orcid.org/0000-0001-5229-9268}{ORCID})

}
\keyword{internal}
