% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower.R
\name{lower}
\alias{lower}
\title{Make a vector of the lower half of a square distance matrix}
\usage{
lower(matrix)
}
\arguments{
\item{matrix}{Square distance matrix with no row names.}
}
\value{
A vector of the lower half of the matrix
}
\description{
Extract lower half of distance matrix
}
\details{
This is a convenience function to obtain the lower half of a matrix
}
\examples{
## Not run:
## m <- matrix(rnorm(25), 5)
## l_m <- lower(m)

## End (Not run)
}
\author{
Bill Peterman \href{mailto:Peterman.73@osu.edu}{Peterman.73@osu.edu}
}
