% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_dist.R
\name{pca_dist}
\alias{pca_dist}
\title{Function to calculate PCA distance between individuals}
\usage{
pca_dist(gi, n_axes = 16, scale = TRUE)
}
\arguments{
\item{gi}{Object of class \code{genind} from the \code{adegenet} package}

\item{n_axes}{Number of principal component axes to retain to calculate pairwise distances}

\item{scale}{(Default = TRUE) Calculated distance will be rescaled 0-1}
}
\description{
Calculate PCA distance
}
\examples{
## Not run:
df <- data.frame(locusA = c("11","11","12","32"),
                locusB = c("37","34","55","15"),
                locusC = c("22","22","21","22"))
df

obj <- adegenet::df2genind(df,
                          ploidy = 2,
                          ncode = 1)

pca_dist(gi = obj,
        n_axes = 4)

pca_dist(gi = obj,
        n_axes = 4,
        scale = FALSE)
## End (Not run)
}
\author{
Bill Peterman \href{mailto:Peterman.73@osu.edu}{Peterman.73@osu.edu}
}
