% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_01.R
\name{scale_01}
\alias{scale_01}
\title{Scale vector}
\usage{
scale_01(x)
}
\arguments{
\item{x}{Vector to be scaled}
}
\description{
Scale vector to range from 0-1
}
\details{
Scale vector of values to range 0-1
}
\examples{
x <- runif(5, -10, 10)
scale_01(x)
}
\author{
Bill Peterman
}
